using Observable.Interval instead, cleaner to read

+++ Droid/Sources/CancellableAlert.cs
@@ -72,7 +72,7 @@ namespace RxTests.Droid
 			return this;
 		}
 
-		public ICancellableAlert SetTimeRemaining (int millisec)
+		public ICancellableAlert SetTimeRemaining (long millisec)
 		{
 			SetOKTitle (string.Format ("OK ({0})", Math.Round (millisec / 1000d)));
 			return this;
